+++ crates/dies-executor/src/control/variable.rs
@@ -7,27 +7,27 @@ pub trait Variable:
     Copy
     + Add<Self, Output = Self>
     + AddAssign
-    + Mul<f32, Output = Self>
+    + Mul<f64, Output = Self>
     + Sub<Output = Self>
-    + Div<f32, Output = Self>
+    + Div<f64, Output = Self>
 {
     /// Returns the zero value for this type.
     fn zero() -> Self;
 }
 
-impl Variable for f32 {
+impl Variable for f64 {
     fn zero() -> Self {
         0.0
     }
 }
 
-impl Variable for Vector2<f32> {
+impl Variable for Vector2<f64> {
     fn zero() -> Self {
         Vector2::zeros()
     }
 }
 
-impl Variable for Vector3<f32> {
+impl Variable for Vector3<f64> {
     fn zero() -> Self {
         Vector3::zeros()
     }
